extracted general iter macros into submodule for organization
